feat: move some func about genesis from tools/gengen/util to pkg/genesis
